Webapp/requirements.txt
torch
torchvision
pillow
pandas
streamlit